separate monitor and auditLogger
